POC for dropping webpack in the css chain
this commit does not build all 4 variations on the css file, only produces cn-tailwind.css right now.

however, it does this with only `yarn build` inside the css directory and using `postcss` direclty to build the css

building the ui-guide with `make build-ui` at the root allows one to serve the ui-guide from the `dist` directory.

the ui-guide mostly works but classes that changed from tailwind 0.6 to 2.x such as `pin-t`, `pin-t-full` etc have to be changed in the ocelot purescript component code now which will be the subject of the following commit