add notes with analysis of classes used before
just-classes.txt contains all the unique class strings that were found in class array definitions…the class-arrays.purs file is the concatenation of all the lines of purescript that directly assigned these classes to arrays of HP.ClassName.

the code-with-classes-embedded is all the code that used / assigned tailwind classes in a more ad hoc fashion, needs to be checked over too.

next these classes have to be vetted against contemporary tailwind css classes - we already know that the `pin-` type classes have to change, there may be others